refactor(hooks): share current branch fallback
