+++ src/hooks/lib/current-branch.ts
@@ -1,0 +1,14 @@
+import { getCurrentBranch } from '../hook-io.js';
+
+export interface CurrentBranchOptions {
+  readBranch?: () => string;
+}
+
+export function currentHookBranch(options: CurrentBranchOptions = {}): string {
+  try {
+    const branch = (options.readBranch ?? getCurrentBranch)().trim();
+    return branch || 'unknown';
+  } catch {
+    return 'unknown';
+  }
+}
