+++ src/hooks/pr-kaizen-clear.ts
@@ -19,7 +19,8 @@ import { join } from 'node:path';
 import { gh } from '../lib/gh-exec.js';
 import { readJsonObjectFile } from '../lib/json-file.js';
 import { type HookInput, readHookInput, writeHookOutput, traceNullInput, traceHookEvent } from './hook-io.js';
-import { appendHookAuditLog, currentHookBranch } from './lib/audit-log.js';
+import { appendHookAuditLog } from './lib/audit-log.js';
+import { currentHookBranch } from './lib/current-branch.js';
 import { formatGateSignal } from './lib/gate-signal.js';
 import { verifyIssueRef, type RefStatus } from './lib/issue-ref-verifier.js';
 import { parseGithubPrUrl } from '../lib/github-pr.js';
